requirements.txt
line-bot-sdk
flask
openai
googletrans==3.1.0a0
